requirements.txt
Flask
Flask-mysql
